requirements.txt
Flask==3.0.0
pymongo==4.6.1
openai==1.6.1
requests==2.31.0
beautifulsoup4==4.12.2
APScheduler==3.10.4
python-dotenv==1.0.0
Werkzeug==3.0.1
lxml==5.1.0
numpy==1.26.3
gunicorn==21.2.0

